+++ domain/Score/AgainstTrait.php
@@ -9,30 +9,30 @@ use SportsHelpers\Against\Side as AgainstSide;
 
 trait AgainstTrait
 {
-    protected int $home;
-    protected int $away;
+    protected float $home;
+    protected float $away;
 
-    public function getHome(): int
+    public function getHome(): float
     {
         return $this->home;
     }
 
-    public function setHome(int $home): void
+    public function setHome(float $home): void
     {
         $this->home = $home;
     }
 
-    public function getAway(): int
+    public function getAway(): float
     {
         return $this->away;
     }
 
-    public function setAway(int $away): void
+    public function setAway(float $away): void
     {
         $this->away = $away;
     }
 
-    public function get(AgainstSide $side): int
+    public function get(AgainstSide $side): float
     {
         return $side === AgainstSide::Home ? $this->getHome() : $this->getAway();
     }
